Move stats printing to main
